% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/opensky-package.R
\docType{package}
\name{opensky-package}
\alias{opensky}
\alias{opensky-package}
\title{opensky: Templates for 'ggplot2' and 'R Markdown'}
\description{
Templates for 'ggplot2' and 'R Markdown'. 'opensky' provides functions to include templates for 'xaringan' slides, 'ggplot2' helpers, and 'R Markdown' documents.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/emptyfield-ds/opensky}
  \item Report bugs at \url{https://github.com/emptyfield-ds/opensky/issues}
}

}
\author{
\strong{Maintainer}: Malcolm Barrett \email{malcolmbarrett@gmail.com} (\href{https://orcid.org/0000-0003-0299-5825}{ORCID})

}
\keyword{internal}
