% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/use_blank_rmd.R
\name{use_blank_rmd}
\alias{use_blank_rmd}
\title{Use a blank R Markdown file with reasonable settings}
\usage{
use_blank_rmd(save_as, open = is_interactive())
}
\arguments{
\item{save_as}{Path of file to create, relative to root of active project.
Defaults to \code{template}}

\item{open}{Open the newly created file for editing? Happens in RStudio, if
applicable, or via \code{\link[utils:file.edit]{utils::file.edit()}} otherwise.}
}
\description{
Use a blank R Markdown file with reasonable settings
}
