% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/use_ggplot2_utils.R
\name{use_ggplot2_utils}
\alias{use_ggplot2_utils}
\title{Use ggplot2 templates}
\usage{
use_ggplot2_utils(save_as = NULL, open = FALSE)
}
\arguments{
\item{save_as}{Path of file to create, relative to root of active project.
Defaults to \code{template}}

\item{open}{Open the newly created file for editing? Happens in RStudio, if
applicable, or via \code{\link[utils:file.edit]{utils::file.edit()}} otherwise.}
}
\description{
Including \code{theme_minimal2()}, \code{geom_point2()}, and other variants
}
