% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/use_r_hub.R
\name{use_r_hub}
\alias{use_r_hub}
\title{Use an ignored \code{rhub.R} file}
\usage{
use_r_hub(save_as = "rhub.R", open = is_interactive())
}
\arguments{
\item{save_as}{Path of file to create, relative to root of active project.
Defaults to \code{template}}

\item{open}{Open the newly created file for editing? Happens in RStudio, if
applicable, or via \code{\link[utils:file.edit]{utils::file.edit()}} otherwise.}
}
\description{
Use an ignored \code{rhub.R} file
}
