% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/use_testthat_setup.R
\name{use_testthat_setup}
\alias{use_testthat_setup}
\title{Create \code{setup.R} in the testthat file}
\usage{
use_testthat_setup(open = is_interactive())
}
\arguments{
\item{open}{Open the newly created file for editing? Happens in RStudio, if
applicable, or via \code{\link[utils:file.edit]{utils::file.edit()}} otherwise.}
}
\description{
Create \code{setup.R} in the testthat file
}
