% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/use_xaringan.R
\name{use_xaringan_kakashi}
\alias{use_xaringan_kakashi}
\alias{use_xaringan_tozan}
\title{Use xaringan templates}
\usage{
use_xaringan_kakashi(save_as, open = is_interactive())

use_xaringan_tozan(save_as)
}
\arguments{
\item{save_as}{Path of file to create, relative to root of active project.
Defaults to \code{template}}

\item{open}{Open the newly created file for editing? Happens in RStudio, if
applicable, or via \code{\link[utils:file.edit]{utils::file.edit()}} otherwise.}
}
\description{
Use xaringan templates
}
