% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/opensky-package.R
\docType{package}
\name{opensky-package}
\alias{opensky}
\alias{opensky-package}
\title{opensky: What the Package Does (One Line, Title Case)}
\description{
What the package does (one paragraph).
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/emptyfield-ds/opensky}
  \item Report bugs at \url{https://github.com/emptyfield-ds/opensky/issues}
}

}
\author{
\strong{Maintainer}: Malcolm Barrett \email{malcolmbarrett@gmail.com} (\href{https://orcid.org/0000-0003-0299-5825}{ORCID})

}
\keyword{internal}
