% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/use_utils.R
\name{use_utils}
\alias{use_utils}
\alias{edit_utils}
\title{Use \code{utils.R} with \verb{\%||\%} and \verb{\%nin\%}}
\usage{
use_utils(save_as = "utils.R", open = is_interactive())

edit_utils(open = is_interactive())
}
\arguments{
\item{save_as}{Path of file to create, relative to root of active project.
Defaults to \code{template}}

\item{open}{Open the newly created file for editing? Happens in RStudio, if
applicable, or via \code{\link[utils:file.edit]{utils::file.edit()}} otherwise.}
}
\description{
Use \code{utils.R} with \verb{\%||\%} and \verb{\%nin\%}
}
